Remove more extension dependencies from tests
