+++ src/shared/vscode/interfaces.ts
@@ -1,5 +1,5 @@
-import { CompletionItemProvider, DebugConfigurationProvider, DebugSession, DebugSessionCustomEvent, DefinitionProvider, ReferenceProvider, RenameProvider, TreeDataProvider, TreeItem, Uri } from "vscode";
-import { Outline } from "../../extension/analysis/analysis_server_types";
+import { CompletionItem, CompletionItemProvider, DebugConfigurationProvider, DebugSession, DebugSessionCustomEvent, DefinitionProvider, ReferenceProvider, RenameProvider, TextDocument, TreeDataProvider, TreeItem, Uri } from "vscode";
+import { AvailableSuggestion, Outline } from "../../extension/analysis/analysis_server_types";
 import { LogCategory, LogSeverity, TestStatus, VersionStatus } from "../enums";
 import { DebugCommandHandler } from "../interfaces";
 import { WorkspaceContext } from "../workspace";
@@ -61,4 +61,19 @@ export interface TestResultsProvider extends TreeDataProvider<TestItemTreeItem> 
 
 export interface TestItemTreeItem extends TreeItem {
 	status: TestStatus;
+}
+
+export interface DelayedCompletionItem extends CompletionItem {
+	autoImportUri: string;
+	document: TextDocument;
+	enableCommitCharacters: boolean;
+	filePath: string;
+	insertArgumentPlaceholders: boolean;
+	nextCharacter: string;
+	offset: number;
+	relevance: number;
+	replacementLength: number;
+	replacementOffset: number;
+	suggestion: AvailableSuggestion;
+	suggestionSetID: number;
 }
